refactor: update new imports for IconCalendarWeek component
