+++ src/presentation/globals/components/icons/Icons.tsx
@@ -1,24 +1,6 @@
 import type { IconProps } from '@/presentation/globals/presentation.types';
 import { IconWrapper } from '@/presentation/globals/components/icons/IconWrapper';
 
-export function IconCalendarWeek(props: IconProps) {
-  return (
-    <IconWrapper {...props}>
-      <path d="M4 7a2 2 0 0 1 2 -2h12a2 2 0 0 1 2 2v12a2 2 0 0 1 -2 2h-12a2 2 0 0 1 -2 -2v-12z" />
-      <path d="M16 3v4" />
-      <path d="M8 3v4" />
-      <path d="M4 11h16" />
-      <path d="M7 14h.013" />
-      <path d="M10.01 14h.005" />
-      <path d="M13.01 14h.005" />
-      <path d="M16.015 14h.005" />
-      <path d="M13.015 17h.005" />
-      <path d="M7.01 17h.005" />
-      <path d="M10.01 17h.005" />
-    </IconWrapper>
-  );
-}
-
 export function IconCircle({
   className = '',
   porcentage = 0,
